chore: extract wrapper and render helpers for test
