--- src/test/utils/queryWrapper.tsx
@@ -1,13 +1,0 @@
-import type { ReactNode } from 'react'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { Suspense } from 'react'
-
-export const queryClient = new QueryClient()
-
-export const queryWrapper = ({ children }: { children: ReactNode }) => {
-  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-}
-
-export const queryWrapperWithSuspense = ({ children }: { children: React.ReactNode }) => (
-  <Suspense fallback={<div>Loading...</div>}>{queryWrapper({ children })}</Suspense>
-)
