+++ src/theme/Flex/Flex.test.tsx
@@ -1,9 +1,7 @@
-import { render, screen } from '@testing-library/react'
+import { screen } from '@testing-library/react'
 import { describe, it, expect } from 'vitest'
-import { ThemeProvider } from 'src/theme/ThemeProvider'
+import { renderWithTheme } from 'src/test/utils/wrapper'
 import { Flex } from './Flex'
-
-const renderWithTheme = (ui: React.ReactNode) => render(<ThemeProvider>{ui}</ThemeProvider>)
 
 describe('Flex', () => {
   describe('children', () => {
