+++ src/theme/View/View.test.tsx
@@ -1,9 +1,7 @@
-import { render, screen } from '@testing-library/react'
+import { screen } from '@testing-library/react'
 import { describe, it, expect } from 'vitest'
-import { ThemeProvider } from 'src/theme/ThemeProvider'
+import { renderWithTheme } from 'src/test/utils/wrapper'
 import { View } from './View'
-
-const renderWithTheme = (ui: React.ReactNode) => render(<ThemeProvider>{ui}</ThemeProvider>)
 
 describe('View', () => {
   describe('children', () => {
